Revert split classes
